+++ R/str_start_end.R
@@ -9,8 +9,6 @@
 #'          be a regular-expression object, as returned by \code{\link[stringr]{regex}}.
 #'          Alternatively, use \code{regex = TRUE} to treat \code{pattern} as a regular
 #'          expression rather than a fixed string.
-#' @param regex Logical, if \code{TRUE}, \code{pattern} is treated as a regular
-#'          expression rather than a fixed string.
 #'
 #' @inheritParams find_var
 #'